Ai_Backed/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
numpy
faiss-cpu
groq
